bug 1505233: move Rule to processor
This moves socorro.lib.transform_rules to socorro.processor.rules.base.
Rules are only used by the processor, so it's easier if the code is
centralized there.
