+++ socorro/processor/rules/memory_report_extraction.py
@@ -2,7 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from socorro.lib.transform_rules import Rule
+from socorro.processor.rules.base import Rule
 
 
 # Kinds of memory reporters.
